Implement Storage Class
