+++ src/seedu/addressbook/storage/Storage.java
@@ -1,0 +1,43 @@
+package seedu.addressbook.storage;
+
+import java.nio.file.Path;
+
+import seedu.addressbook.data.AddressBook;
+import seedu.addressbook.data.exception.IllegalValueException;
+
+public abstract class Storage {
+    
+    /** Default file path used if the user doesn't provide the file name. */
+    public static final String DEFAULT_STORAGE_FILEPATH = "addressbook.txt";
+
+    /**
+     * Signals that the given file path does not fulfill the storage filepath constraints.
+     */
+    public static class InvalidStorageFilePathException extends IllegalValueException {
+        public InvalidStorageFilePathException(String message) {
+            super(message);
+        }
+    }
+
+    /**
+     * Signals that some error has occured while trying to convert and read/write data between the application
+     * and the storage file.
+     */
+    public static class StorageOperationException extends Exception {
+        public StorageOperationException(String message) {
+            super(message);
+        }
+    }
+    
+    /**
+     * Returns true if the given path is acceptable as a storage file.
+     * The file path is considered acceptable if it ends with '.txt'
+     */
+    protected static boolean isValidPath(Path filePath) {
+        return filePath.toString().endsWith(".txt");
+    }
+    
+    public abstract AddressBook load() throws StorageOperationException;
+    public abstract void save(AddressBook book) throws StorageOperationException;
+    public abstract String getPath();
+}
